Added IO_FLAG_FS_DATA flags, and fixed page cache deadlocks.
The IO_FLAG_FS_DATA flag allows the file system to inform IO that this
read/write is on behalf of a file. The IO subsystem can use that to
properly determine whether or not to link physical memory between the disk
and the file.

Also used the IO_FLAG_FS_DATA flags to avoid infinite recursion in
cleaning a very dirty page cache.

Also fixed a deadlock in the dirty file objects list where a file object
that was being flushed caused a new file object to become dirty and need
to get added to the list.