Improved dt and bc debugger commands.
The dt debug command now does a symbol search if it finds a structure type
with size zero. This may find the wrong symbol, but is generally better than
displaying no data.

The bc debug command now removes a breakpoint from the list even if it fails
to be cleared. This is useful for removing stale breakpoints for unloaded
modules. A warning is sent to the user.