Stricter variable assignments in Chalk.
This change introduces the concept of null to the Chalk interpreter. It
also requires that variables be assigned before use (whereas before they
were just created if they did not exist). Also added the get function,
which returns the value tied to a key in a dictionary or null if no value
for that key was found.