Avoid aggressively cleaning in page cache thread.
Avoid immediately going back to clean if there's something dirty, and
limit to the five second timer. The problem is that when the disk cache
gets big, the page cache thread chews up CPU time looking for a couple of
dirty pages. It also holds the disk lock shared, which blocks out things
like tar from making much progress.

In the future, consider adding dirty page cache entries to a dirty list in
the file object. The list entry is already in the page cache entry and is
currently unused when the entry is dirty. To flush/clean entries, pick an
entry off the list, then scan the tree starting there for the largest
contiguous run of dirtyness that can be achieved.