Converted netcore queued locks to shared exclusive.
Now that the shared exclusive locks are theoretically cheaper to acquire,
use them to protect networking resources that are rarely modified, but often
search. The plugin and link locks are great candidates because all socket
create and bind requests have to acquire at least one of them, but those
only need to be shared acquires.