Added network device reference counting.
The contract between netcore and a physical layer networking device driver
is a two way street. The device driver needs to call into netcore upon packet
reception and netcore needs to invoke the device's driver to send packets.
The device's driver holds netcore in memory by linking against it, so as long
as the device is around, netcore will be around. Prior to this change, netcore
had no references on the device or the device's driver, despite having a
function table and context belonging to the device.

This change allows for safe device removal by having netcore take a reference
on the device when its link is added to netcore. Netcore can then share that
reference with sockets that bind to the link. When the device receives a
removal IRP, it now notifies netcore of its removal, but cannot yet destroy
its context. Netcore owns the link and still has a handle to the driver's
interface and context. The driver must wait until netcore calls back into
the driver via the DestroyLink interface function and then it can safely
destroy the physical device's context. Netcore will no longer call into the
device's driver after calling DestroyLink.