Added the netlink network with generic protocol.
This change introduces the base netlink network. It is a local network in
which sockets are not backed by the physical links. This allows for socket-
based communication between kernel mode and user mode. Netcore needed a
little updating to handle such link-less sockets and to accommodate the
fact that all kernel-owned netlink sockets are assigned port 0. No other
existing networks let a socket bind to port 0.

This change also adds the basics for the generic netlink protocol. Subtypes
need to be added to allow kernel side processing, but that is up to the
individual component (e.g. 802.11) to register for and implement.