Wire netlink sockets up to user mode.
This change includes bug fixes on the kernel side and the introduction of a
basic netlink.h. It is located at netlink/netlink.h in the libc headers for
now. As netlink sockets are not part of POSIX, this may move depending on
which packages come to depend on netlink sockets.