Put dirty page cache entries on a list.
When flushing the disk, the page cache thread spends a lot of time
iterating through the tree looking for dirty entries. This change puts
dirty entries on an unordered list, so that flush can get to them quickly.
Once flush grabs an entry from the list, it still creates a run via tree
iteration, to get maximal contiguous runs.

Also release and reacquire the file object lock during flush on the page
cache if there is contention. This allows user-mode writes to continue to
come in while the page cache is busy flushing a dirty disk.