Fixed bugs in DMA controllers.
This change switches the EDMA and Broadcom DMA controllers to allocate one
transfer per channel. Access to each channel is synchronized by the DMA core
and, as a result, only one transfer is ever necessary for the device specific
controllers. As a side effect, this also removes a call to allocate from
non-paged pool while at dispatch. This could be bad news for the DMA transfer
if it triggers pool expansion (the DMA would have failed) and could cause a
deadlock if the paging thread hit the pool expansion.

This change also fixes the EDMA and Broadcom controllers to pump the next
transfer through the channel if a transfer is returned by the DMA core
completion routine.