Fixed DMA transfer cancel bug.
This change fixes a bug in the DMA cancel logic. DmaCancel will cancel the
currently executing transfer and immediately submit any transfers that are
pending on that channel. But, even with a successful cancel, the old
transfer's completion interrupt may have been in mid-flight. The DMA device
controllers (EDMA and Broadcom) previously could not tell the difference
between an old interrupt for the cancelled transfer and an interrupt for the
newly submitted transfer. This could have resulted in the new transfer being
terminated early. The fix is to have the DMA device controllers check to
make sure the channel is not active before completing a transfer.