Fix for missing ACPI tables on HP dv9000.
The tables were in an area marked as ACPI NV storage, which is an area of
memory previously skipped when mapping ACPI tables.

Also added in the RSDP search region for a segmented address in the EBDA,
since that's how it's described in some documentation. The dv9000 doesn't
actually need this.