Added BIOS text mode support to base video.
The HP dv9000 laptop has a dying video card (in fact part of a class
action suit that resulted in a recall, but which I learned of too late).
When the video card is just sputtering enough to display something, it
enumerates zero modes via VESA, but displays the default text mode
properly. Support the old text mode for future oddball scenarios.