Fix lost dirty page cache entries.
If the flush page cache entries loop exited early, entries on the local
list would be lost, remaining in a destroyed local list and never getting
cleaned.

Also made an attempt to map IO buffer pages one at a time if VA space is
too fragmented for the combined approach.