Page cache bug fixes.
1. Added a "timid" page cache flush to address a problem where the page
cache was getting too large while FAT was zero-filling a file. The code
before didn't do a page cache trim when the FS_DATA flag was set because
it was worried it would try to flush the same file that's being written to
further up the execution stack. Since that lock is already held, doing
this would hang the thread. So the page cache got so huge it swallowed all
of memory, which then locked things up in other ways.

This change introduces a "timid" mode of flushing page cache entries,
which gives up after one attempt of trying to acquire a file object lock
while flushing. This way easy entries (which is almost all of them) can be
trimmed while not having to stress about the one file that's inaccessible.

2. Fixed some accounting issues with the mapped and mapped dirty page
counts.

3. Added a necessary empty list check in remove page cache entries from
list.

4. Fixed a bug with how file objects were being marked dirty that resulted
in dirty file objects not ending up on the dirty file objects list.