Fix a null dereference in the page cache.
When flushing page cache entries, the page cache thread tries to be
nice by dropping and reacquiring the file object lock if the file object
is contended. It also sometimes traverses the tree in that loop. If the
lock is dropped and the node that loop is using is ripped out of the tree,
then that loop will fault trying to get the next tree node. Fix this by
not dropping the lock if traversing via the tree.