Fix page cache crash in mmap_io_shared perf test.
MmpDestroyImageSectionMappings may go to mark a page cache entry dirty.
IoMarkPageCacheEntryDirty uses the the owner flag to determine whether
to use the backing entry or not, all before acquiring the file object
lock. IopLinkPageCacheEntries changes the flags and then sets the backing
entry. If these things are going on at the same time, a null
dereference can occur. Modify mark page cache entry dirty a bit to
avoid this situation.