Minor perf improvement to page cache IO contention.
While the page cache thread is flushing entries, allow the lock to be
released/reacquired even if traversing the tree by explicitly looking to
see if the node was removed and not traversing the tree if so. This allows
the lock to be released and reacquired more often in the page cache
thread, so user mode isn't so backed up behind it.