Fix illegal VFP instruction on ARM.
When cloning a thread, the current thread needs to raise to dispatch level
to ensure it doesn't get swapped out and lose the FPU while it's trying to
save the FPU state to the new cloned thread. This is not an issue on x86
since all FPU registers are volatile across function calls (like fork), so
no FPU state needs to be copied to the new thread.