Moved network device info registration to netcore.
Every networking device needs to respond to network device information
requests. As the actual requests are handled by the networking core, this
change moves the registration/de-registration into netcore as well. This is
made possibly by a recent change that passes the OS device into netcore on
link add as part of the properties.