Record RSSI for BSS entries.
When a packet is received, the physical layer driver should be able to
determine the signal strength and pass this information to the 802.11 core.
It is then up to the 802.11 core to store it as necessary. This change adds
the simplest method of updating RSSI data (i.e. doing it after every packet).