Fixed corruption in sh.
ShEscapeSpecialCharacters was processing the entire output, including the
last character which was assumed to be a null terminator, but may not
actually have been. If this turned out to be a character that required
escaping, then the resulting string would end in an escape character,
since that last character in the resulting string that should have been a
null is now a character that has been escaped. In a command substitution
this could cause the next character after the substitution to get treated
literally, if that next character was a control character. So control
characters were showing up in the output, which is bad.

The fix is to not process that assumed null terminator. For safety also
actually terminate the resulting string if one was created.