Fix race in IoWaitForIoObjectState.
IO_OBJECT_STATE is not synchronized in the case of pipes. Make was trying
to write to a pipe, and when the wait in IoWaitForIoObjectState finished,
the Events observed were still zero. This caused the pipe code to exit
with STATUS_BROKEN_PIPE, which killed make.

The fix is to loop until there are some valid event bits to return. The
poll code was not changed because it was assumed callers can handle
returns from poll where no events turned out to be ready.