Bring back the flush lock.
The system was rebooting inappropriately with dirty data still in the
cache. This is because flush page cache entries pulls all the dirty
entries onto a local list to work with them, so the IoFlush calls saw
nothing to do. The fix can't easily be done with the file object lock
because the page cache thread drops the file object lock during block
device flushes, so holding the lock exclusive doesn't necessarily mean
there aren't things on a local dirty list. A global flush lock
synchronizes explicit flushes with each other and non-explicit flushes
like the page cache thread.