Replaced SOCKET_NETWORK with NET_DOMAIN_TYPE.
The SOCKET_NETWORK enum was no longer a good descriptor for the types it
enumerated. ARP, EAPOL, Ethernet, and 802.11 are not socket level networks.
But it is difficult to break the namespace apart as there are multiple
overlapping subsets within that share properties. For example, socket types
like IPv4 and data link types like Ethernet both have their own address
formats but low level protocols like ARP do not. The Venn diagram of these
subsets gets complicated. So, with a simple rename to the more generic
'domain' type, the enum is now less socket specific but keeps all network
types in the same namespace.