Aligned AF_* with NetDomain* values.
This change allows the C-library to directly pass the supplied socket domain
value to the OsSocket* APIs. The goal of this change is to easily allow new
socket domains to be added without C-library modifications. The newly added
user mode values are directly passed to the netcore driver making it easy
for a third-party to add the user mode and driver components without changes
to the core Minoca system and libraries.