Added domain validation on socket creation.
Not all NET_DOMAIN_TYPE values are valid for sockets. As these values are
passed directly from user mode to Netcore, it needs to validate the value
before searching for a NET_NETWORK_ENTRY.

As part of this, now that not all network entries can be found during socket
creation, this change loosens the netcore network registration requirements.
Non-socket networks do not need to provide the socket-level interface
callbacks such as bind, listen, accept, and send.