Fixed getsockopt to silently truncate the data.
The POSIX specification indicates that if the getsockopt option value is
larger than the supplied buffer, getsockopt should silently truncate the
data and return successfully. This change honors that behavior, but does not
perpetuate it. The Minoca OS interface will return the truncated data but
indicate that the supplied buffer was too small.

This change also includes various bug fixes for the socket options including
removing some duplicate code, allowing network protocols to override the
default socket option behaviors, range checking on C library integer options
(as they are converted to ULONGs by the OS), and other clean-up.