Fixed bugs in Unix sockets.
The first bug, found by accident, caused a deadlock in IopUnixSocketConnect.
If a stream Unix socket tried to connect to itself, the code attempted to
acquire a lock it already held.

The second bug was in converting from a NETWORK_ADDRESS to a sockaddr_un.
The code calling ClpConvertFromNetworkAddress made some assumptions that if
the given sockaddr's sun_path was already correct, that it would remain
correct after the call. This was not the case as the routine fixed up a stack
allocated sockaddr_un and copied that over the supplied sockaddr, scrozzing
any path that may have been.