Added C library type conversion interfaces.
In order to avoid educating the C library about how to convert between all
C library style types and native Minoca OS types, this changes adds a way
for a library or application to register a conversion callback. The first
consumer of this will be the netlink library. This will allow the library
and netlink consumers to use the C library sockets API. Said API takes
sockaddr_nl addresses and calls the native libminocaos sockets API which
takes NETWORK_ADDRESS addresses. Thus the need for a conversion callback.