Fixed user mode memory corruption in kernel.
When an image section is split in two by an unmap, its dirty and inherited
bitmaps are split as well, potentially with a shift. The loop that set up
those new bitmaps was sometimes not copying the second portion of the last
word of the new bitmap. When paging was enabled, this meant that heap
pages that were paged out before the split were getting simply zeroed
rather than paged back in when faulted in. Compiling GCC with 512MB of RAM
usually hit this.