Added kernel netlink ACK and NACK responses.
If a netlink message requests an ACK or an error occurs while handling a
message, then the kernel should send an ACK or a NACK. This change adds this
behavior to the kernel. In order to make it easy on all netlink socket
protocols, this error handling is done by the netlink network module in a
code path common to all netlink protocols.