Fixed message-based sockets receive behavior.
Message-based sockets are supposed to receive one and only one message at
a time. They do not adhere to the SO_RCVLOWAT socket option or the
MSG_WAITALL flag (to use the C library names).

This change also fixes a smaller bug where sockets are supposed to fail
shutdown requests if they are not connected.