Improvements to mbgen.
The top level project file can now specify the default build file name and
source root relative to the directory the project file is in. Removed
those settings from the command line since they're specific to the project
layout and not configurable.

Also created the Makefile/build.ninja regeneration rules so that mbgen
gets rerun if a build script changes.