Refactored elf.c to also compile 64-bit.
elf.c now relies on a set of defines that point to different but
analagous structures and rename the functions defined in the file.
This allows 64-bit ELF support to simply be a recompile of the 32-bit ELF
support with one extra define flipped on.

Functions that do not change between 32 and 64 bit (including that they do
not call a specialized routine) are refactored to elfcomm.c.

For now the top level Im functions do not have the 64 bit functions wired
up.