Added netlink set info for joining multicast groups.
This change wires setsockopt up to netlink core's get/set information to
allow users to add and remove sockets from multicast groups. It also
includes a few bug fixes to the netlink core to allow a socket to repeatedly
bind to new multicast groups as long as the provided port ID is the same.