Added 802.11 netlink get scan results response.
The goal of this change is to get 802.11 core to respond to the get scan
results netlink request command. Said response requires packaging a
multi-part netlink message (terminated with the DONE message) and sending
that back to the caller as one packet. As such, helper routines netlink
offered to send commands needed some updating. The updates include routines
to append the generic and base netlink headers and attributes to a packet.
This makes it easy for a consumer to build a message starting from the
header. Send then rewinds the packet's data offset and sends off the message.