Added opt-out of sequence checking to libnetlink.
In preparation for adding scan support to netcon, this change adds an
optional netlink library socket flag to disable the automatic netlink
sequence number inclusion and validation.

Another part of this change is to clean up the netlink library API to match
the kernel's netlink message building helpers a bit more. One day, a
common library could be shared between libnetlink and netcore.