Added wireless network scan support to netcon.
This change adds support to the netcon utility for querying available
wireless networks and displaying the results. The bulk of this change was
to build out support in the netlink library for handling multi-part
messages. NlReceiveMessage not takes in a set of parameters, one of which
is a callback routine that is invoked for each protocol layer message. The
receiving application can use this callback and a private context to store
the results of the protocol specific messages.