Added more 802.11 scan results.
This change updates the 802.11 core to return more data when the netlink
scan results message is received. This required a bit of rework to save
the entire information element received by a 802.11 mgmt probe response.

The change also updates the netcon utility to make use of the newly reported
data.