Allowed BSS entries with less secure ciphers.
Before this change, if a BSS was not using CCMP for both group and pairwise
ciphers, the 802.11 driver would not save it. It turns out that most
networks will support CCMP for pairwise encryption (as it is the strongest)
and some lesser protocol (TKIP) for group encryption as it is a lower
common denominator for nodes in the network. So, this change stops tossing
BSS's just because they use TKIP for group encryption.