Trim old BSS entries before scans.
Scans can occur when user mode requests the most up-to-date list of
available networks or when use mode is attempting to join a network. In
either case it is not desireable to show old BSS entries; they may no longer
be in range. This change defines the BSS aging timeout at 10 seconds.