Added 802.11 scanning for associated stations.
While a station is associated with an AP, it is actively listening on a
specific channel and likely filtering incoming packets backed on the BSSID
of the AP. This is at odds with an attempt to scan for available wireless
networks. A scan needs to send probe requests out over each channel and
receive responses from various BSS's (all with unique BSSIDs).

In order to do this, this change updates the scan thread code to handle
pausing data traffic before briefly switching to another channel to scan
and then resuming said traffic after jumping back to the original channel.