Fixed Debugger UI preferences for minimized windows.
When windows are minimized, they actually get some non-zero height and
width and a seemingly bogus X,Y offset. The debugger UI preferences load
code was trying to protect against minimization by not reloading preferences
for windows with 0 width or 0 height. Unfortunatley, that's doesn't work.

This change tracks the size of the window with each move in a global
variable, but ignores the size when a minimize or maximize action occur. The
window size is always taken from this global in case the window is currently
minimized or maximized.