Updated MSG_TRUNC behavior to return real length.
If MSG_TRUNC is supplied to a socket receive call (for data gram and raw
sockets), then it will return the real length of the received packet even
if it is truncated. This is useful in combination with MSG_PEEK to determine
the right size of buffer to allocate for a receive.