Fixed disabling profiling when a core is idle.
In order to make sure that all cores are done using the profiling buffers,
the disable code waits for each core to see another clock interrupt. Idle
cores may never see a clock interrupt. This fix sends a clock IPI to wake
up all the cores so that the profiling buffers can be safely destroyed.