Removed the kernel terminal's command history.
The kernel was saving command history when operating in canonical mode.
While convenient in some cases, it is insecure in others. In particular, if
a program asked for a user's password while in canonical mode (e.g. getpass)
that password was stored in the kernel's command history and could be
retrieved by running 'cat' and hitting the up arrow key.