Fixed a clock interrupt storm on Galileo/Quark.
Turns out the major source of the laggy terminal and awful perf was due to
a clock interrupt storm. The Galileo was taking tens of thousands
of clock interrupts per second while not doing anything. It turns out that
the way we were setting one-shot timers in the local APIC was causing them
to fire immediately. Disable the timer briefly while programming to avoid
this.