Fixed BeagleBone Black hang.
The BeagleBone Black was hanging coming out of an aborted standby when it
was trying to reset the M3. The M3 simply wouldn't respond. I believe this
is because the mailbox interrupt was disabled (albiet briefly) when the A8
pulsed the mailbox interrupt. Normally the pulsing behavior is required,
since the M3 cannot quiesce its own mailbox interrupt, and if it reruns
the interrupt handler with an invalid command it will no longer be able to
complete a sleep transition later. But for aborted power transitions the
A8 and M3 might be racing, so leave the interrupt hammering the M3 until
it's known to have responded.