Fixed CM3 hang in BeagleBone Black.
Added updated Cortex M3 firmware for the BeagleBone Black, and map the
M3's data region for debugging purposes. The problem in the firmware was
that the M3 was getting the interrupt for the A8 WFI, but if the A8 woke
back up before the M3 could take it down, it left the mailbox interrupt
disabled and never got an MPU_WAKE interrupt. So the A8 was unable to
communicate with it.