Updated netcon to not required a device ID.
For the typical case where a system only has one wireless networking device,
this makes it easier on the user by automatically finding the device ID
when a wireless command is specified. If there is more than one WLAN device
present, the netcon utility will simply print an ambiguity notice and list
the available devices.