Updated UEFI firmware to support Raspberry Pi 3.
This change allows the Raspberry Pi 2 image (rpi2.img) to boot the Raspberry
Pi 3 in 32-bit mode. The major differences between the RPI2 and RPI3 in
32-bit mode are that the UART is not assigned GPIO pins 14 and 15 at boot,
the max APB clock rate is 400 MHz, and the MPIDR values start at 0 for the
cores rather than 0xF00.

This change also upgrades the Raspberry Pi firmware blobs for all versions
of the Raspberry Pi.