Added shutdown to the Raspberry Pi 1 & 2.
The BCM2709 based chips do not support an official way to shutdown via the
power registers. The Raspberry Pi firmware, however, looks for a special
value in the reset status register. If found, the firmware will not proceed
with the next boot and turn off most of the chip. It's the best that can
be done for these platforms.