Reduced KD's connection wait time to 2 seconds.
Now that the host and target both send connection synchronization bytes,
the 10 second timeout is excessive. It used to be necessary because the
target had to catch the host's connection request which was sent every 5
seconds - a big enough window was needed to catch one. Now that the host
will send a connection request only once the target announces itself, there
is no need for the large window.