Rearranged versioning out of the Makefile (mostly).
Up until now we've used command line flags like
-DSYSTEM_BUILD_TIME=... and -DSYSTEM_VERSION_MAJOR=1
on the compile commands in the Makefile. This won't work for the new build
system because 1) it causes the command line used to compile to change
every second and 2) Ninja won't go run shell commands not related to
updating a target.

Now version information is generated into a version.h file in the obj
directory, which sources include. A shell script generates the
version.h file. I've wired up version.h to depend on .git/HEAD so
that whenever the commit changes the version.h files are rebuilt. The
slight tradeoff here is that now .git/HEAD needs to be there, even for
out-of-version-control builds (like what the automation does). Seems like
a worthy tradeoff right now to always have up to date commit numbers in dev
builds.

REVISION is currently broken, and will always report as 0. There are too
many things using it, so I'm doing it in a separate change (coming the
change after next).