Nailed down the REVISION.
What was REVISION is now VERSION_SERIAL, and is only used by files
including a generated version.h. This removes -DREVISION from the GCC
command line. Unfortunately a lot of apps used it, and those apps for
which it wasn't worth it to add the build infrastructure for their own
version.h no longer contain REVISION or its new equivalent.

Also dealt with the FIRMWARE_VERSION_* definitions in the UEFI platforms.
Many now take their information from uefi/core's version.h information.
The BIOS date string is not quite in the right format, but I don't know if
anyone cares. For those platforms (Panda and RPi) that don't have fixed
MAC addresses, continue to hardcode in their own versions (which have been
revved here) to avoid the node IDs changing all the time in the
automation.