Added specific target options to mingen.
If arguments are specified on the command line, mingen can now create
build files that build only those targets. Additionally, the -g option
tells mingen not to create a generator rule to recreate the build.ninja or
Makefile.

These options will be useful when the bootstrapping build files are
created, as we will need to check in the build files necesary for building
mingen itself.