Short term tuning attempt at fixing VA fragmentation.
The C201 Chromebook gets itself into a situation where VA is fragmented
enough to trigger a warning, but there's really not much allocated. The
build then grinds to a near halt because the page cache is constantly
trying to unmap things. Most of the fixed allocations seem to come from
two places: the kernel's MDL descriptors themselves, and IO buffers which
I think are pinned because of the FAT window cache.

This change attempts to mitigate the fragmentation by 1) allocating larger
chunks for MDLs and FAT windows and 2) Trying to keep cluster allocations
towards the beginning of the disk.

To truly handle this problem, we're going to need to make these fixed
allocations less fixed. Probably many of those MDL pages can be handed
back, and some (but not all) of the FAT cache can also be dropped if
things get tight virtually. Perhaps moving the kernel virtual allocator to
a bitmap would be even better.