Fix for new Raspberry Pi SMBIOS names.
The product names in Raspberry Pi now have things like Model B Rev 1.1 in
them. Rather than matching all of those, msetup now just matches the
prefix of what it has in the SMBIOS name.

This means that Raspberry Pi 2 had to be moved above Raspberry Pi, so that
it matched the more specific 2 first. 1 was added to the original
Raspberry Pi, but there are also models like "Raspberry Pi Compute" and
"Raspberry Pi Zero" that cannot match the 1 string.