Updated AM335x CM3 firmware for sleep hang.
This will hopefully fix an occasional issue on the BeagleBone Black where
the system would idle out and never come back. I believe the issue was
that the M3 was clearing all interrupts in its WFI interrupt handler, but
if the wake interrupt were already there at that time, it would be lost.