More RK32 SD changes.
The RockChip SD controller still craps out. I noticed the TRM has some
obscure language describing that the DWC divisor register can only be set
to 0 or 1. We seemed to be setting it to many different values. It
appeared to be sticking, but I wonder if it was only ever setting a
divisor of 1 or 2 (off of 99MHz).

This change sets the divisor in the CRU, and only uses a value of 0 or 1
in the Designware divisor register. It was a little tricky to override the
DesignWare function on UEFI. I needed to expose some structures previously
internal to SD DWC to Veyron.

We'll see if this makes a difference.