Fixed noisy BCM2709 interrupts and timers.
The Broadcom interrupt controller baked into the Raspberry Pi appears to
execute posted writes (the writes succeed before completion). This, in
addition to interrupt priority being implemented in software, meant that
some interrupts were firing twice. The second instance would occur as soon
as the first instance re-enabled interrupts to call the ISR, despite the
first begin interrupt call disabling the interrupt. The solution is to
always read a register after writing it. Heavy, but necessary.

While investigating this fix, it was noticed that the Boardcom ARM timer was
generating interrupts only for the interrupt controller to find no pending
interrupt. This was due to a race between the ARM timer interrupt being
acknowledged by one core and being re-armed by another core. The re-arming
core was clearing the interrupt state before the acknowledging core had a
chance to observe the interrupt. The fix is to not clear the interrupt state
when arming the timer.