Reduced clock timer hardware arm requests.
On multi-core systems without per-processor clock timers, the HL module
determines the correct clock setting by looking at all processors. The
algorithm was being conservative to make sure to always set the next
one-shot deadline, even if it was already set in the hardware. This resulted
in unnecessary calls to program the hardware in two cases. The first was
while all the cores were on their way to idle. Any core going idle after the
one-shot deadline was set, set it again with a legitimate future value. The
second came after the timer expired. The one-shot deadline remains valid
until the processor that owns it got to run and clear it. But before that
happened, other cores were waking first and setting the timer to one-shot
with 0 clock ticks as they saw a deadline in the past. Seemed wasteful.

The fix is to just record the last one-shot deadline set. If the next core
determines that the same deadline is the next deadline, it will just leave
the timer as-is. This solves both of the above cases.