Add LALR test slash new Chalk parser.
This is currently in a bit of a work-in-progress state, as the only real
goal was to validate that the LALR parser was correctly parsing the
grammar.

ckmatch.y is the same as ck.y except the tokens are converted to literals
so yacc will sequence the tokens correctly. ckmatch2.y has the rules
rearranged so the non-terminal symbols will sequence correctly. With these
changes the LALR grammar table matches yacc exactly, and correctly parses
the test.ck file.