Rotated ck grammar node order to make more sense.
Now that matching with yacc is no longer important, uncomment grammar
nodes that were ordered in the more logical way. Yacc would define symbols
as soon as they were seen, which resulted in this sort of interposed
waterfall pattern. It didn't make much sense logically.

Also fixed a bug in the parser resulting from the fact that the final
rule is no longer the first rule.