Removed hardware module kernel services interface.
This change removes the limited API exposed to the hardware modules. The
modules are now allowed to directly call kernel APIs. The original goal of
the limited API was to sandbox the hardware modules, knowing that they would
eventually be converted to dynamic libraries and that they should not depend
on any kernel APIs outside of HL, RTL, and arch. They are the lowest point
on any callstack and should not take dependencies on higher level
routines. But as new hardware modules require more HL routines and can
take advantage of additional RTL routines, it finally makes sense to break
apart the API.

In the new world, hardware modules will include kernel.h. The responsibility
of choosing the correct APIs to call is on the developer. Creating a more
limited header file with just routines for the hardware modules was
considered, but that would need to include hl.h which sets off a chain of
header file dependencies. Most of kernel.h would have been required. So, the
plan is to include kernel.h and hope people don't shoot themselves in the
foot.