Fixed module init loops to be consistent.
Some files were looping until a module array entry was found to be NULL.
Others were calculating the array length and using a precise for-loop. This
changes switches them all to be consistent, opting for the more space
efficient for-loop (i.e. the arrays don't need an extra NULL entry).