Fixed BCM2709 system timer reads.
This change updates the read routine for the two 32-bit system timers and
the 64-bit system time counter. Before the 64-bit counter was advertised as
a 32-bit counter and the 32-bit timers were reading back the time remaining,
assuming they had been armed for periodic intervals. The read routine is
really supposed to return the timer value for free-running modes.