Fixed potential torn read in HlpTimerExtendedQuery.
Through code inspection, it was noticed that the read of *CurrentCountPointer
in HlpTimerExtendedQuery could be torn. The same routine will update that
memory location atomically whenever a rollover is detected. That update
could race with another core (or lower runlevel on the same core) attempting
to read the 64-bit current count, resulting in the tear.

This tear is bad because the reported time may end up being one rollover in
the past or one rollover in the future. The incorrect value will never get
set in the timer's current count (as it will always fail the atomic compare
exchange).