Added one-shot mode for BCM2709 system timers.
The BCM2709 system timers interrupt based on a compare register. When the
counter equals that compare register, the timer interrupts and does not
reprogram the compare register for more periodic interrupts. Before this
change, periodic mode was being faked by storing the period in the timer
structure and when acknowledge interrupt saw a non-zero period, it would
rearm the timer. Disarm would set the period to zero. This was simple enough
and the worst case was the wrong period was programmed.

That worst case, however, could be bad if a hard deadline were missed. So
this change introduces a one-shot mode and improves the rearming algorithm.
To do so, a generation counter is added to make sure acknowledge interrupt
does not get a torn state between the period and mode set on arm. Acknowledge
interrupt will be very careful to make sure it arms the timer correctly,
handling any races with other calls to arm or disarm.