Fix for Ps handle table locking.
The conditions for whether or not a handle table lock needed to be
acquired were insufficient in the case where a multithreaded process
dropped back down to a single thread. This caused a thread to acquire the
lock but never release it, and ultimately die with the lock still held.

The new code defers creation of the lock until things go
multithreaded, which should speed things up a tiny bit for single threaded
apps. Once the lock is created, it is always acquired.