Kernel terminal fixes.
1. Be aware of the fact that the terminal settings might change within one
invocation of a terminal IO function if it blocked on the events. This was
causing test failures in expect because the spawned cat process wasn't
picking up the raw terminal settings.
2. More fixups for making sure the event flags are set correctly for the
terminal. Spurious exits from poll/select were causing expect test
failures.
3. Remove the WorkingInputLock and simply combine with the InputLock. It
was starting to get hairy when you had to acknowledge that the terminal
settings might change mid-function, and the simpler single lock is
probably better anyway.